fix: install supabase js deps and setup lib on server
