+++ .env.example
@@ -1,4 +1,6 @@
-SUPABASE_DATABASE_URL=your-database-url
+SUPABASE_DATABASE_URL=your-supabase-database-url
+SUPABASE_URL=your-supabase-url
+SUPABASE_ANON_KEY=your-supabase-anon-key
 
 ACCESS_TOKEN_SECRET=your-access-token-secret
 REFRESH_TOKEN_SECRET=your-refresh-token-secret
