+++ env.d.ts
@@ -5,6 +5,8 @@ declare global {
       SUPABASE_DATABASE_URL: string
       ACCESS_TOKEN_SECRET: string
       REFRESH_TOKEN_SECRET: string
+      SUPABASE_URL: string
+      SUPABASE_ANON_KEY: string
     }
   }
 }
