+++ package.json
@@ -20,6 +20,7 @@
     "@radix-ui/react-label": "^2.0.2",
     "@radix-ui/react-select": "^2.0.0",
     "@radix-ui/react-slot": "^1.0.2",
+    "@supabase/supabase-js": "^2.44.2",
     "@tanstack/react-table": "^8.17.3",
     "bcrypt": "^5.1.1",
     "class-variance-authority": "^0.7.0",
