+++ server/lib/supabase.ts
@@ -1,0 +1,5 @@
+import { createClient } from '@supabase/supabase-js'
+
+const supabase = createClient(process.env.SUPABASE_URL, process.env.SUPABASE_ANON_KEY)
+
+export { supabase }
